+++ bang_py/game_manager.py
@@ -14,16 +14,9 @@ from .event_logic import EventLogicMixin
 from .card_handlers import CardHandlersMixin
 from .general_store import GeneralStoreMixin
 from .turn_phases import TurnPhasesMixin
-from .characters.jesse_jones import JesseJones
-from .characters.jose_delgado import JoseDelgado
-from .characters.kit_carlson import KitCarlson
-from .characters.pat_brennan import PatBrennan
-from .characters.pedro_ramirez import PedroRamirez
 from .characters.vera_custer import VeraCuster
 from .cards.bang import BangCard
 from .cards.missed import MissedCard
-from .cards.panic import PanicCard
-from .cards.jail import JailCard
 from .cards.general_store import GeneralStoreCard
 
 
@@ -208,20 +201,6 @@ class GameManager(EventLogicMixin, CardHandlersMixin, GeneralStoreMixin, TurnPha
         """Select which character a player will use. Defaults to the first."""
         return options[0]
 
-    def apply_new_identity(self, player: Player) -> None:
-        """Swap ``player`` to their unused character if the new identity event is active."""
-
-        if not self.event_flags.get("new_identity"):
-            return
-        new_char = player.metadata.unused_character
-        if not new_char:
-            return
-        old_char = player.character
-        player.character = new_char
-        player.metadata.unused_character = old_char
-        player.reset_stats()
-        player.character.ability(self, player)
-        player.health = min(2, player.max_health)
 
     def _deal_roles_and_characters(self) -> None:
         role_deck = self._build_role_deck()
@@ -242,385 +221,8 @@ class GameManager(EventLogicMixin, CardHandlersMixin, GeneralStoreMixin, TurnPha
             player.character.ability(self, player)
             player.metadata.game = self
 
-    def _handle_equipment_start(self, player: Player) -> bool:
-        """Process start-of-turn equipment effects.
-
-        Reactivates green equipment, resolves Dynamite and Jail, and returns
-        ``True`` if the player's turn should continue.
-        """
-        self._reactivate_green_equipment(player)
-        if not self._resolve_dynamite(player):
-            return False
-        return self._process_jail(player)
-
-    # ------------------------------------------------------------------
-    # Equipment helpers
-    def _reactivate_green_equipment(self, player: Player) -> None:
-        """Refresh green equipment and reapply modifiers."""
-        for eq in list(player.equipment.values()):
-            if eq.card_type == "green" and not getattr(eq, "active", True):
-                eq.active = True
-                modifier = int(getattr(eq, "max_health_modifier", 0))
-                if modifier:
-                    player._apply_health_modifier(modifier)
-
-    def _resolve_dynamite(self, player: Player) -> bool:
-        """Handle Dynamite at turn start. Returns ``False`` if the player dies."""
-        dyn = player.equipment.get("Dynamite")
-        if dyn and getattr(dyn, "check_dynamite", None):
-            next_idx = self.turn_order[(self.current_turn + 1) % len(self.turn_order)]
-            next_player = self._players[next_idx]
-            exploded = dyn.check_dynamite(player, next_player, self.deck)
-            if exploded:
-                self.discard_pile.append(dyn)
-                self.on_player_damaged(player)
-                if not player.is_alive():
-                    self._begin_turn()
-                    return False
-        return True
-
-    def _process_jail(self, player: Player) -> bool:
-        """Resolve Jail effects and return ``False`` if the turn is skipped."""
-        jail = player.equipment.get("Jail")
-        if not jail:
-            return True
-        if self.event_flags.get("no_jail"):
-            player.unequip("Jail")
-            self.discard_pile.append(jail)
-            return True
-        if getattr(jail, "check_turn", None):
-            skipped = jail.check_turn(player, self.deck)
-            self.discard_pile.append(jail)
-            if skipped:
-                self.current_turn = (self.current_turn + 1) % len(self.turn_order)
-                self._begin_turn()
-                return False
-        return True
-
-    def _handle_character_draw_abilities(self, player: Player) -> bool:
-        """Trigger characters that modify the draw phase.
-
-        Returns ``True`` if the standard draw and play phases should be skipped.
-        """
-        ability_chars = (
-            JesseJones,
-            KitCarlson,
-            PedroRamirez,
-            JoseDelgado,
-            PatBrennan,
-        )
-
-        if isinstance(player.character, ability_chars):
-            player.metadata.awaiting_draw = True
-            for cb in self.turn_started_listeners:
-                cb(player)
-            return True
-        return False
-
-    def _begin_turn(self, *, blood_target: Player | None = None) -> None:
-        if not self.turn_order:
-            return
-        self.current_turn %= len(self.turn_order)
-        idx = self.turn_order[self.current_turn]
-        player = self._players[idx]
-        self.phase = "draw"
-        self.reset_turn_flags(player)
-
-        player = self._run_start_turn_checks(player)
-        if not player:
-            return
-
-        self.draw_phase(player, blood_target=blood_target)
-        self.play_phase(player)
-        player.metadata.bangs_played = 0
-        for cb in self.turn_started_listeners:
-            cb(player)
-
-    def _run_start_turn_checks(self, player: Player) -> Player | None:
-        """Apply start-of-turn effects returning the acting player or ``None``."""
-        player = self._apply_event_start_effects(player)
-        if not player:
-            return None
-        if self._maybe_revive_ghost_town(player):
-            return None
-        if not self._handle_equipment_start(player):
-            return None
-        if self._handle_character_draw_abilities(player):
-            return None
-        return player
-
-    def end_turn(self) -> None:
-        """Finish the current player's turn and advance to the next."""
-        if not self.turn_order:
-            return
-        idx = self.turn_order[self.current_turn]
-        player = self._players[idx]
-        self.phase = "discard"
-        self.discard_phase(player)
-        self.event_flags.pop("turn_suit", None)
-        self._reset_green_equipment(player)
-        if self._handle_vendetta(player):
-            return
-        self._finish_ghost_town(player)
-        self._advance_turn()
-
-    # ------------------------------------------------------------------
-    # Turn end helpers
-    def _reset_green_equipment(self, player: Player) -> None:
-        """Reactivate green equipment at the end of the turn."""
-        for eq in list(player.equipment.values()):
-            if eq.card_type == "green" and not getattr(eq, "active", True):
-                eq.active = True
-                modifier = int(getattr(eq, "max_health_modifier", 0))
-                if modifier:
-                    player._apply_health_modifier(modifier)
 
 
-    def _advance_turn(self) -> None:
-        """Move the turn pointer and start the next turn."""
-        if self.event_flags.get("reverse_turn"):
-            self.current_turn = (self.current_turn - 1) % len(self.turn_order)
-        else:
-            self.current_turn = (self.current_turn + 1) % len(self.turn_order)
-        self._begin_turn()
-
-
-    def _pre_card_checks(
-        self, player: Player, card: BaseCard, target: Optional[Player]
-    ) -> bool:
-        return (
-            self._card_in_hand(player, card)
-            and self._run_card_play_checks(player, card, target)
-            and self._check_target_restrictions(player, card, target)
-            and self._check_event_restrictions(player, card)
-        )
-
-    def _card_in_hand(self, player: Player, card: BaseCard) -> bool:
-        """Return ``True`` if ``card`` is currently in ``player``'s hand."""
-        return card in player.hand
-
-    def _run_card_play_checks(
-        self, player: Player, card: BaseCard, target: Optional[Player]
-    ) -> bool:
-        """Execute registered pre-play checks."""
-        for cb in self.card_play_checks:
-            if not cb(player, card, target):
-                return False
-        return True
-
-    def _check_target_restrictions(
-        self, player: Player, card: BaseCard, target: Optional[Player]
-    ) -> bool:
-        """Validate distance and target based limitations."""
-        return (
-            self._bang_target_valid(player, card, target)
-            and self._panic_target_valid(player, card, target)
-            and self._jail_target_valid(card, target)
-        )
-
-    # ------------------------------------------------------------------
-    # Target restriction helpers
-    def _bang_target_valid(
-        self, player: Player, card: BaseCard, target: Optional[Player]
-    ) -> bool:
-        """Check range and sniper restrictions for Bang! cards."""
-        if not (isinstance(card, BangCard) and target):
-            return True
-        if player.distance_to(target) > player.attack_range:
-            return False
-        if self.event_flags.get("sniper") and player.metadata.use_sniper:
-            bang_count = sum(isinstance(c, BangCard) for c in player.hand)
-            return bang_count >= 2
-        return True
-
-    def _panic_target_valid(
-        self, player: Player, card: BaseCard, target: Optional[Player]
-    ) -> bool:
-        """Panic can only target players within distance 1."""
-        if isinstance(card, PanicCard) and target:
-            return player.distance_to(target) <= 1
-        return True
-
-    def _jail_target_valid(self, card: BaseCard, target: Optional[Player]) -> bool:
-        """Jail cannot be played on the sheriff."""
-        if isinstance(card, JailCard) and target and isinstance(target.role, SheriffRoleCard):
-            return False
-        return True
-
-    def _check_event_restrictions(
-        self, player: Player, card: BaseCard
-    ) -> bool:
-        """Check event related card play restrictions."""
-        return not (
-            self._jail_blocked(card)
-            or self._judge_blocked(card)
-            or self._handcuffs_blocked(player, card)
-        )
-
-    def _jail_blocked(self, card: BaseCard) -> bool:
-        """Return ``True`` if Jail cards are currently banned."""
-        return self.event_flags.get("no_jail") and isinstance(card, JailCard)
-
-    def _judge_blocked(self, card: BaseCard) -> bool:
-        """Return ``True`` if blue cards are disallowed by The Judge."""
-        return self.event_flags.get("judge") and card.card_type in {"blue", "green"}
-
-    def _handcuffs_blocked(self, player: Player, card: BaseCard) -> bool:
-        """Return ``True`` if Handcuffs restricts ``player`` from playing ``card``."""
-        if not self.event_flags.get("handcuffs") or not self.event_flags.get("turn_suit"):
-            return False
-        active = self._players[self.turn_order[self.current_turn]]
-        return player is active and getattr(card, "suit", None) != self.event_flags["turn_suit"]
-
-    def _is_bang(self, player: Player, card: BaseCard, target: Optional[Player]) -> bool:
-        return isinstance(card, BangCard) or (
-            player.metadata.play_missed_as_bang and isinstance(card, MissedCard) and target
-        )
-
-    def _can_play_bang(self, player: Player) -> bool:
-        if self.event_flags.get("no_bang"):
-            return False
-        count = player.metadata.bangs_played
-        gun = player.equipment.get("Gun")
-        extra = player.metadata.doc_free_bang
-        unlimited = (
-            player.metadata.unlimited_bang
-            or getattr(gun, "unlimited_bang", False)
-            or extra > 0
-        )
-        limit = self.event_flags.get("bang_limit", 1)
-        return count < limit or unlimited
-
-
-
-    def play_card(self, player: Player, card: BaseCard, target: Optional[Player] = None) -> None:
-        """Play ``card`` from ``player`` against ``target`` if allowed."""
-        if not self._pre_card_checks(player, card, target):
-            return
-
-        is_bang = self._is_bang(player, card, target)
-        if is_bang and not self._can_play_bang(player):
-            return
-
-        player.hand.remove(card)
-        before = target.health if target else None
-        self._dispatch_play(player, card, target)
-        self._notify_card_played(player, card, target)
-        self._apply_post_play(player, card, target, before, is_bang)
-
-    def _notify_card_played(self, player: Player, card: BaseCard, target: Optional[Player]) -> None:
-        """Call registered card played listeners."""
-        for cb in self.card_played_listeners:
-            cb(player, card, target)
-
-    def _apply_post_play(
-        self,
-        player: Player,
-        card: BaseCard,
-        target: Optional[Player],
-        before: Optional[int],
-        is_bang: bool,
-    ) -> None:
-        """Handle damage, discard and Bang! bookkeeping."""
-        self._apply_damage_and_healing(player, target, before)
-        self._pass_left_or_discard(player, card)
-        if is_bang:
-            self._update_bang_counters(player)
-        self._draw_if_empty(player)
-
-    def _apply_damage_and_healing(
-        self, source: Player, target: Optional[Player], before: Optional[int]
-    ) -> None:
-        """Trigger damage or heal callbacks if ``target`` changed health."""
-        if target and before is not None:
-            if target.health < before:
-                self.on_player_damaged(target, source)
-            elif target.health > before:
-                self.on_player_healed(target)
-
-    def _draw_if_empty(self, player: Player) -> None:
-        """Draw a card if the player has an empty hand and may draw."""
-        if player.metadata.draw_when_empty and not player.hand:
-            self.draw_card(player)
-
-    def _update_bang_counters(self, player: Player) -> None:
-        """Update Bang! play counters for ``player``."""
-        if player.metadata.doc_free_bang > 0:
-            player.metadata.doc_free_bang -= 1
-        else:
-            player.metadata.bangs_played += 1
-
-    def discard_card(self, player: Player, card: BaseCard) -> None:
-        """Discard ``card`` from ``player`` and process event effects."""
-        if card in player.hand:
-            player.hand.remove(card)
-            self._pass_left_or_discard(player, card)
-            if not self.event_flags.get("river"):
-                handle_out_of_turn_discard(self, player, card)
-            if player.metadata.draw_when_empty and not player.hand:
-                self.draw_card(player)
-
-    def sid_ketchum_ability(self, player: Player, indices: List[int] | None = None) -> None:
-        """Discard two cards to heal ``player`` for one life."""
-        if len(player.hand) < 2 or player.health >= player.max_health:
-            return
-        discard_indices = indices or list(range(2))
-        discard_indices = sorted(discard_indices, reverse=True)[:2]
-        for idx in discard_indices:
-            if 0 <= idx < len(player.hand):
-                card = player.hand.pop(idx)
-                self._pass_left_or_discard(player, card)
-        player.heal(1)
-        self.on_player_healed(player)
-
-    def _auto_miss(self, target: Player) -> bool:
-        if not self._should_use_auto_miss(target):
-            return False
-        if self._use_miss_card(target):
-            return True
-        if self._use_bang_as_miss(target):
-            return True
-        if self._use_any_card_as_miss(target):
-            return True
-        return False
-
-    def _should_use_auto_miss(self, target: Player) -> bool:
-        """Return ``True`` if automatic Missed! can be used."""
-        if self.event_flags.get("no_missed"):
-            return False
-        return target.metadata.auto_miss is not False
-
-    def _discard_and_record(self, player: Player, card: BaseCard) -> None:
-        self._pass_left_or_discard(player, card)
-        if not self.event_flags.get("river"):
-            handle_out_of_turn_discard(self, player, card)
-
-    def _use_miss_card(self, target: Player) -> bool:
-        miss = next((c for c in target.hand if isinstance(c, MissedCard)), None)
-        if miss:
-            target.hand.remove(miss)
-            self._discard_and_record(target, miss)
-            target.metadata.dodged = True
-            return True
-        return False
-
-    def _use_bang_as_miss(self, target: Player) -> bool:
-        if target.metadata.bang_as_missed:
-            bang = next((c for c in target.hand if isinstance(c, BangCard)), None)
-            if bang:
-                target.hand.remove(bang)
-                self._discard_and_record(target, bang)
-                target.metadata.dodged = True
-                return True
-        return False
-
-    def _use_any_card_as_miss(self, target: Player) -> bool:
-        if target.metadata.any_card_as_missed and target.hand:
-            card = target.hand.pop()
-            self._discard_and_record(target, card)
-            target.metadata.dodged = True
-            return True
-        return False
 
     def chuck_wengam_ability(self, player: Player) -> None:
         """Lose 1 life to draw 2 cards, usable multiple times per turn."""
